02-LoL-ETL/requirements.txt
requests
pandas
python-dotenv
loguru